+++ src/libs/constants/index.ts
@@ -1,1 +1,1 @@
-import "./fullPageObjArray";
+export *  from "./fullPageObjArray";
