--- src/libs/hocks/fullPage-intersectionObserverr.ts
@@ -1,17 +1,0 @@
-// const observer = new IntersectionObserver(
-//   entries => {
-//     entries.forEach(entry => {
-//       if (entry.isIntersecting) {
-//         // 페이지가 뷰포트에 들어왔을 때 수행할 동작
-//         entry.target.classList.add("active");
-//       } else {
-//         // 페이지가 뷰포트에서 나갔을 때 수행할 동작
-//         entry.target.classList.remove("active");
-//       }
-//     });
-//   },
-//   {
-//     root: null, // 뷰포트
-//     threshold: 0.5 // 페이지가 뷰포트에 50% 이상 보이면 콜백 실행
-//   }
-// );
