+++ src/libs/hooks/index.ts
@@ -1,0 +1,1 @@
+export * as default  from './useFullPage'
